Add Dockerfile and config for docker usage
Dockerfile that installs the nfl-notifier application and its
dependencies. The Dockerfile copies credentials.json and token.pickle.
These files contain the login information for the googleapiclient.
The config_docker.toml is also copied to the docker
container and renamed to config.toml. It specifies the logging path
(/log/nfl_notifier.log).